LuNge/requirements.txt
flask==3.0.3 
python-dotenv==1.0.1 
groq==0.13.1
deep_translator==1.11.4 
pymongo==4.9.1
motor==3.6.0
cachetools==5.5.0 
hypercorn==0.17.3 
asgiref==3.8.1 
aiohttp==3.10.5
uuid==1.30
